+++ changes.md
@@ -4,6 +4,9 @@ Change Log
 ### changes in 0.2
 
 * changed loadBundle() from MarkerSupport to be readBundle() in Neo4jFileLoader
+* added an Option[Logger] to readBundle()
+* pass an implicit Logger to all that need it
+
 
 ### changes in 0.1